fixing conflict again

+++ src/main/java/com/grille/dao/SkillRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by jiawei on 16/05/2017.
  */
-public interface SkillRepository extends JpaRepository<Skill, Long> {
+public interface SkillRepository extends JpaRepository<Skill, Integer> {
 
 }
